fd-commands: add pldm_completion_code
The results for the corresponding pldm requests have an u8 field
called CompletionCode. These completion codes are typically a
set of PLDM_BASE_CODES and a selection of FW_UPDATE codes.
To pass these return codes in a generic and type safe way, we
want to define enums that implement the `From` trait for u8,
individually for each response.

Signed-off-by: leongross <leon.gross@9elements.com>